Completed following items: Forest takes too long to deplete Lower forest count Add number of hits taken/dealt Newbie protection is needed Anti-stuck check The red circle for report Descriptions for assigning labors coordination on top of minimap should be map centered need to add description to forest rate Ranking page(Attack should say Attack Point, etc to avoid confusion Client is not adding extra cost when a city is in battle Cancel action should ask for confirmation Need to describe garrison notification of being attacked moving speed formula revisit Current round should be shown even if user cant watch the battle client always get stuck at loading Receive defense point when stationed Star rating Structure downgrade action

+++ Game/Battle/BattleFormulas.cs
@@ -4,7 +4,6 @@ using System;
 using Game.Data;
 using Game.Data.Stats;
 using Game.Setup;
-using Game.Logic.Conditons;
 using Game.Data.Troop;
 
 #endregion
@@ -127,7 +126,7 @@ namespace Game.Battle {
         }
 
         internal static ushort GetStaminaReinforced(City city, ushort stamina, uint round) {
-            return (ushort) (stamina);
+            return stamina;
         }
 
         internal static ushort GetStaminaRoundEnded(City city, ushort stamina, uint round) {
